backend: Major refactor

+++ src-tauri/src/db.rs
@@ -1,13 +1,9 @@
 use crate::config::Category;
 
 use super::Config;
-use dirs;
+use log::{debug, info, warn};
 use rusqlite::{params, Connection, Result};
 use std::path::PathBuf;
-
-struct File {
-    path: String,
-}
 
 #[derive(Debug)]
 struct CategoryEntry {
@@ -22,41 +18,58 @@ impl PartialEq for CategoryEntry {
 }
 
 pub struct Database {
-    connection: Connection,
+    connection: Option<Connection>,
 }
 
 impl Database {
-    pub fn open() -> Database {
-        let con = Connection::open(Database::db_path()).expect("Failed to open database");
-        let db = Database { connection: con };
-        db.create_tables().expect("Failed to generate tables");
-        db
+    pub fn new() -> Self {
+        Database { connection: None }
     }
 
-    fn db_path() -> PathBuf {
-        if let Some(config_dir) = dirs::config_dir() {
-            config_dir.join("files").join("files.sqlite")
-        } else {
-            panic!("Config directory not found")
+    pub fn open(&mut self, path: PathBuf) -> Result<()> {
+        info!("Opening database at {0}", path.display());
+        self.connection = match Connection::open(path) {
+            Ok(con) => Some(con),
+            Err(err) => {
+                return Err(err);
+            }
+        };
+        if self.is_initialized() {
+            self.create_tables().expect("Failed to generate tables");
         }
+        Ok(())
+    }
+
+    pub fn is_initialized(&self) -> bool {
+        self.connection.is_some()
     }
 
     fn get_tags(&self) -> Result<Vec<String>> {
-        self.connection
-            .prepare("SELECT tag FROM tags")
-            .expect("Tag select failed")
-            .query_map([], |row| row.get(0))?
-            .collect::<Result<Vec<String>>>()
+        if let Some(ref con) = self.connection {
+            con.prepare("SELECT tag FROM tags")
+                .expect("Tag select failed")
+                .query_map([], |row| row.get(0))?
+                .collect::<Result<Vec<String>>>()
+        } else {
+            // TODO not very clean (has to be chanaged eveywhere)
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn delete_tag(&self, tag: &String) -> Result<usize> {
-        self.connection
-            .execute("DELETE FROM tags WHERE tag = ?1", params![tag])
+        if let Some(ref con) = self.connection {
+            con.execute("DELETE FROM tags WHERE tag = ?1", params![tag])
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn insert_tag(&self, tag: &String) -> Result<usize> {
-        self.connection
-            .execute("INSERT INTO tags(tag) VALUES (?1)", params![tag])
+        if let Some(ref con) = self.connection {
+            con.execute("INSERT INTO tags(tag) VALUES (?1)", params![tag])
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn tag_consistency(&self, config_tags: &Vec<String>) -> Result<()> {
@@ -65,7 +78,7 @@ impl Database {
         // Delete unused tags
         for tag in tags.iter() {
             if !config_tags.contains(tag) {
-                println!("Deleting tag {}", tag);
+                debug!("Deleting tag {}", tag);
                 self.delete_tag(tag)?;
             }
         }
@@ -73,7 +86,7 @@ impl Database {
         // Insert new tags
         for tag in config_tags.iter() {
             if !tags.contains(tag) {
-                println!("Inserting tag {}", tag);
+                debug!("Inserting tag {}", tag);
                 self.insert_tag(tag)?;
             }
         }
@@ -81,47 +94,68 @@ impl Database {
     }
 
     fn get_categories(&self) -> Result<Vec<CategoryEntry>> {
-        self.connection
-            .prepare("SELECT id, name FROM categories")?
-            .query_map([], |row| {
-                Ok(CategoryEntry {
-                    id: row.get(0)?,
-                    name: row.get(1)?,
-                })
-            })?
-            .collect::<Result<Vec<CategoryEntry>>>()
+        if let Some(ref con) = self.connection {
+            con.prepare("SELECT id, name FROM categories")?
+                .query_map([], |row| {
+                    Ok(CategoryEntry {
+                        id: row.get(0)?,
+                        name: row.get(1)?,
+                    })
+                })?
+                .collect::<Result<Vec<CategoryEntry>>>()
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn get_category_id(&self, name: &String) -> Result<u32> {
-        self.connection.query_row(
-            "SELECT id FROM categories WHERE name = ?1",
-            params![name],
-            |r| r.get(0),
-        )
+        if let Some(ref con) = self.connection {
+            con.query_row(
+                "SELECT id FROM categories WHERE name = ?1",
+                params![name],
+                |r| r.get(0),
+            )
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn delete_category(&self, id: u32) -> Result<usize> {
-        self.connection
-            .execute("DELETE FROM categories WHERE id = ?1", params![id])
+        if let Some(ref con) = self.connection {
+            con.execute("DELETE FROM categories WHERE id = ?1", params![id])
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn insert_category(&self, name: &String) -> Result<usize> {
-        self.connection
-            .execute("INSERT INTO categories(name) VALUES (?1)", params![name])
+        if let Some(ref con) = self.connection {
+            con.execute("INSERT INTO categories(name) VALUES (?1)", params![name])
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn delete_value(&self, category_id: u32, value: &String) -> Result<usize> {
-        self.connection.execute(
-            "DELETE FROM categoryValue WHERE value = ?1 and category_id = ?2",
-            params![value, category_id],
-        )
+        if let Some(ref con) = self.connection {
+            con.execute(
+                "DELETE FROM categoryValue WHERE value = ?1 and category_id = ?2",
+                params![value, category_id],
+            )
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn insert_value(&self, category_id: u32, value: &String) -> Result<usize> {
-        self.connection.execute(
-            "INSERT INTO categoryValue(category_id, value) VALUES (?1, ?2)",
-            params![category_id, value],
-        )
+        if let Some(ref con) = self.connection {
+            con.execute(
+                "INSERT INTO categoryValue(category_id, value) VALUES (?1, ?2)",
+                params![category_id, value],
+            )
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 
     fn category_consistency(&self, config_cats: &Vec<Category>) -> Result<()> {
@@ -129,47 +163,49 @@ impl Database {
         for category in categories.iter() {
             // Delete unused categories
             if !config_cats.iter().any(|c| c.name.eq(&category.name)) {
-                println!("Removing category {}", category.name);
+                debug!("Removing category {}", category.name);
                 self.delete_category(category.id)?;
             }
         }
 
         // Insert new categories
         for config_category in config_cats {
-            println!("Checking category {}", config_category.name);
+            debug!("Checking category {}", config_category.name);
             if !categories.iter().any(|c| c.name.eq(&config_category.name)) {
-                println!("Inserting category {}", config_category.name);
+                debug!("Inserting category {}", config_category.name);
                 self.insert_category(&config_category.name)?;
 
                 let id: u32 = self.get_category_id(&config_category.name)?;
 
                 for value in config_category.values.iter() {
-                    println!("Inserting value {}", value);
+                    debug!("Inserting value {}", value);
                     self.insert_value(id, value)?;
                 }
             } else {
                 // Here we can be sure that the category in the config also exists
                 // in the database. We need to check if the values match
-                let mut stmt = self.connection.prepare("SELECT categoryValue.value FROM categoryValue JOIN categories ON categoryValue.category_id = categories.id WHERE categories.name = ?1")?;
+                if let Some(ref con) = self.connection {
+                    let mut stmt = con.prepare("SELECT categoryValue.value FROM categoryValue JOIN categories ON categoryValue.category_id = categories.id WHERE categories.name = ?1")?;
 
-                // These are the values of the current category
-                let values: Vec<String> = stmt
-                    .query_map(params![config_category.name], |row| Ok(row.get(0)?))?
-                    .collect::<Result<Vec<String>>>()?;
+                    // These are the values of the current category
+                    let values: Vec<String> = stmt
+                        .query_map(params![config_category.name], |row| Ok(row.get(0)?))?
+                        .collect::<Result<Vec<String>>>()?;
 
-                let id: u32 = self.get_category_id(&config_category.name)?;
+                    let id: u32 = self.get_category_id(&config_category.name)?;
 
-                // Check if value needs to be deleted
-                for val in values.iter() {
-                    if !config_category.values.iter().any(|c| c.eq(val)) {
-                        println!("Removing value {}", val);
-                        self.delete_value(id, val)?;
+                    // Check if value needs to be deleted
+                    for val in values.iter() {
+                        if !config_category.values.iter().any(|c| c.eq(val)) {
+                            debug!("Removing value {}", val);
+                            self.delete_value(id, val)?;
+                        }
                     }
-                }
-                for config_val in config_category.values.iter() {
-                    if !values.iter().any(|c| c.eq(config_val)) {
-                        println!("Inserting value {}", config_val);
-                        self.insert_value(id, config_val)?;
+                    for config_val in config_category.values.iter() {
+                        if !values.iter().any(|c| c.eq(config_val)) {
+                            debug!("Inserting value {}", config_val);
+                            self.insert_value(id, config_val)?;
+                        }
                     }
                 }
             }
@@ -179,15 +215,16 @@ impl Database {
     }
 
     pub fn check_config_consistency(&self, config: &Config) -> Result<(), String> {
+        info!("Performing consistency check");
         let _ = self.tag_consistency(&config.tags).map_err(|err| {
-            println!("Updating tags failed: {err}");
+            warn!("Updating tags failed: {err}");
             format!("Failed to update tags: {err}").to_string()
         });
 
         let _ = self
             .category_consistency(&config.categories)
             .map_err(|err| {
-                println!("Updating cateogires failed: {err}");
+                warn!("Updating cateogires failed: {err}");
                 format!("Failed to update categories: {err}").to_string()
             });
 
@@ -235,7 +272,11 @@ CREATE TABLE IF NOT EXISTS fileValues (
   CONSTRAINT fk_file FOREIGN KEY (file_id) REFERENCES files(id) ON UPDATE CASCADE ON DELETE CASCADE,
   CONSTRAINT fk_value FOREIGN KEY (value_id) REFERENCES categoryValue(id) ON UPDATE CASCADE ON DELETE CASCADE
 );";
-        self.connection.execute_batch(qry)
+        if let Some(ref con) = self.connection {
+            con.execute_batch(qry)
+        } else {
+            Err(rusqlite::Error::InvalidQuery)
+        }
     }
 }
